got rid of the distinction between canonical and non-caonical operators.
It was uncessary so far. The right way to add it to add another function Opid -> bool in a super typeclass of the term signature,
and not to define Opid as a sum type, wich exposes all of the development to this unnecessary detail, and also make the library hard to explain.