WIP: relating the complex DB substitition (with index reduction) to the simpler one.
The difference matters when multiple substititions have to be
performed. There is a choice of always doing {0:=v} which works with
index reduction. This is used in Certicoq.

The other one, which is simpler, does {n:=} with n decreasing.
The current proof of DB to named conversion (subst commutes)
can only be used with the this version of DB substitution for terms
binding multiple vars.