ported to Coq 8.9. String now needs to be imported to use string
notations. But when importing that, it conflicts with list
definitions (e.g. length) and notations.
When Coq.Lists.List is imported, it hides the same-named and same-typed
lemmas in list.v (they were added probably when such items where not
on Coq stdlib): but the variable names in quantifiers is different:
e.g. in_map_iff, which can cause many proofs to break.

The fix strategy was as follows:
Import String.
Import Coq.Lists.List
Import SquiggleEq.list.